revised custom previous it + describe definitions
