+++ tests/fixtures.ts
@@ -43,7 +43,7 @@ export const TEST_APPLICATION_URI = process.env.APPLICATION_URI;
 export const DEMO_APPLICATION_URI = 'https://12341234-1234-1234-1234-123412341234-us-west-2.apps.astra-dev.datastax.com';
 export const ENVIRONMENT = (process.env.APPLICATION_ENVIRONMENT ?? 'astra') as DataAPIEnvironment;
 
-export const initTestObjects = async (preferHttp2 = USE_HTTP2, environment: typeof ENVIRONMENT = ENVIRONMENT): [DataAPIClient, Db, Collection] => {
+export const initTestObjects = (preferHttp2 = USE_HTTP2, environment: typeof ENVIRONMENT = ENVIRONMENT): [DataAPIClient, Db, Collection] => {
   const client = new DataAPIClient(TEST_APPLICATION_TOKEN, {
     httpOptions: { preferHttp2, client: HTTP_CLIENT_TYPE },
     dbOptions: { namespace: DEFAULT_NAMESPACE },
