--- tests/unit/testing.test.ts
@@ -1,15 +1,0 @@
-import { it } from '@/tests/setup';
-
-describe('my.tests', () => {
-  it`long dev`('sfasdfsafdsadf', function () {
-  });
-
-  it`long dev`('sfasdfsafdsadf', async function () {
-  });
-
-  it('sfasdfsafdsadf', function () {
-  });
-
-  it('sfasdfsafdsadf', async function () {
-  });
-});
